#180 Separate examples
