+++ .github/workflows/go-test-windows.yaml
@@ -35,13 +35,13 @@ jobs:
         run: copy testdata/senzing-license/g2.lic "C:\Program Files\Senzing\g2\etc\g2.lic"
       - name: Copy test database files
         run: mkdir "C:\Temp\sqlite" && copy testdata/sqlite/G2C.db "C:\Temp\sqlite\G2C.db"
-      # - name: Run go test
-      #   run: go test -v -p 1 ./...
-      - name: Run go test - main
-        run: go test -v -p 1 ./.
-      - name: Run go test - g2configserver
-        run: go test -v -p 1 ./g2configserver
-      - name: Run go test - g2configmgrserver
-        run: go test -v -p 1 ./g2configmgrserver
-      - name: Run go test - g2productserver
-        run: go test -v -p 1 ./g2productserver
+      - name: Run go test
+        run: go test -v -p 1 ./...
+      # - name: Run go test - main
+      #   run: go test -v -p 1 ./.
+      # - name: Run go test - g2configserver
+      #   run: go test -v -p 1 ./g2configserver
+      # - name: Run go test - g2configmgrserver
+      #   run: go test -v -p 1 ./g2configmgrserver
+      # - name: Run go test - g2productserver
+      #   run: go test -v -p 1 ./g2productserver
